Removes BitcoindCall trait
