+++ rpc/src/main/scala/org/bitcoins/rpc/client/BlockchainRpc.scala
@@ -12,7 +12,7 @@ import scala.concurrent.Future
 /**
   * RPC calls related to querying the state of the blockchain
   */
-protected trait BlockchainRpc extends Client with BitcoindCall {
+protected trait BlockchainRpc extends Client {
   def getBestBlockHash: Future[DoubleSha256Digest] = {
     bitcoindCall[DoubleSha256Digest]("getbestblockhash")
   }
